strlist: introduce `strlist`
As part of teaching `grln` to chose a different key (or column) for
grouping, let's introduce a new library that will manage a list of
string, or `strlist`.

The `strlist` library will provide all mechanism to init(), free(),
get() and add() strings (char *) with the ability to dynamic grow the
allocated-memory. Moreover, the caller can decide to pre-allocate the
`strlist` in order to improve the allocation performance. This value
will be used to grow the list, as well. For example, if the `strlist`
is pre-allocated with `3` items, every time it needs, it will also
grow by 3 items.

The function `split_line` is moved to a `strlist_split()`, which will
split a line into n-terms and return in `strlist` structure.

The following functions are introduced:

  strlist_init()  - Memory-allocated the values and initialize them
  strlist_free()  - Free any allocated item
  strlist_add()   - Add a new item to the list
  strlist_get()   - Get a new item from the list
  strlist_split() - Split a string and return each term in a `strlist`