grln: add the abiliy to chose which column for grouping
`grln` will always chose the first column, for grouping lines
together. However, a lot of times the user wants to pick a different
key for grouping.

Hence, let's teach `grln` to pick a different key by introducing a new
option, namely: `-c <key>` | `--column <key>`, which will instruct
`grln` to chose a different term and, therefore; hash other values.

In order to avoid confusion, negative number are not allowed and any
attempt to use will result in an application error. Furthermore, the
value 0 will be mapped to the 1st column and a WARN message will be
emitted to notify the user the value 1 is used instead.