strlist: remove strlist_get out-of-bound access
strlist_get() will emit an out-of-bound error to the stdout, however;
with the introduction of `-c | --column` option, the user can specify
a greater value then available columns, which makes the program failed
with the aforementioned out-of-bound message.

Instead, let's just print the lines without any color and prevent the
error message. For that, strlist_get() will return NULL when
out-of-bound access is attempted. The caller, then; only checks if the
return is NULL and print the current in-flight line.