grln: add default values to command options
In 6c3e1f0 (strlist: remove strlist_get out-of-bound access,
in 2023-03-23), grln learned how to specify a different char for its
field delimiter. However, this introduces a BUG, as it now there's no
default value for the delimiter, which ends it up been 0 (char's
initial value). Basically, breaking previous behavior of splitting
the fields using a single space.

So, let's ensure that all options values are initialized with sane
defaults, this is done by adding a new function called
`options_defaults()`. To maintain the backward-comparability the
delimiter default value will be a single space, or SP.