grln: teach `grln -d | --delimiter` to accept value of any size > 0
By using `-d | --delimiter` option, the user can specify a different
character other than single SP (" ") to use as field separator.
Sometimes, however, the text uses a longer delimiter which is not
supported. The program even notifies that this is not possible when a
longer delimiter is attempted:

     $ echo hello::world | grln -d ::
     delimiter '::' is not valid, it should be only one char

So, let's teach `grln -d | --delimiter` to accept a delimiter that has
any size bigger than 0. If an empty delimiter is added, we shall
fallback to the default value, which remains a single SP.

In order to accomplish any-size delimiter, the strlist_split() needs
to be refactor to learn how to compare each consecutive sub-string for
properly splitting the terms. Although the new refactored-code, will
maintain the same O(n) complexity, given that we still requires a
single loop to visit each char. The comparison, on the other hand,
will be costlier, given that we now compared n bytes ahead of the
current positions. A representation of the algorithm can be seen like
so:

    ex #1: hello---world
           ^
  	   compare(hel, ---)  # delimiter = ---

    ex #2: hello---world
            ^
	    compare(ell, ---)

    [ ... a few executions later ]

    ex 6#: hello---world
                ^
	        compare(---, ---)  # we found our delimiter

To maintain the same behavior as before, we need to ensure white-space
chars are normalized, thus; a new function, namely: str_isqual_norm(),
is introduced to perform the comparison but treating all white-space
chars as equal.

Finally, two other helper functions are added to:

  * strsub()   - Returns a sub string given a initial position and
                 size of the sub string.
  * is_blank() - Returns true if an string only contains white-space
                 characters, otherwise; returns false.