grln: improve -c | --column option error handling
The `-c | --column` uses `atoi()` function to convert the command-line
string to int. This function, however; does not treat any errors like
when non-numeric value is defined or when the values underflows and/or
overflows.

Therefore, let's refactor `-c | --column` option code to better handle
the aforementioned errors. For that, we need to use the `strtol()`
function that checks whether the value overflow/underflow or if an
invalid representation of the number is given. This error is set in
`errno` value, which is handled by our new code.

Furthermore, in order to make the code lean, the error handling is
wrapped into a new wrapper, namelly: xstrtol(); making the code that
handles the `-c | --column` option simpler.

Also, let's add a `break` that was missing on the `-d | --delimiter`
option :).